Changes to price data generation

+++ backend/backend/models/analyse_pris.py
@@ -6,8 +6,8 @@ from django.utils.timezone import now
 from .analyse_type import AnalyseType
 
 class AnalysePris(models.Model):
-    intern_pris = models.IntegerField(default=0)
-    ekstern_pris = models.IntegerField(default=0)
+    intern_pris = models.FloatField(default=0)
+    ekstern_pris = models.FloatField(default=0)
     gyldig_fra = models.DateTimeField(default=now)
     gyldig_til = models.DateTimeField(blank=True, null=True)
     analyse_type = models.ForeignKey(
